mctl/pyproject.toml
[project]
name = "mctl"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Lykasov",email = "alex@lykasov.com"}
]
readme = "../README.md"
requires-python = ">=3.9"
dependencies = [
    "typer (>=0.20.0,<0.21.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "mctl", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
